tdf#96505 Deleted suffix long (L) where it is possible
Reviewed-on: https://gerrit.libreoffice.org/46041
Tested-by: Jenkins <ci@libreoffice.org>

+++ sw/source/core/fields/flddat.cxx
@@ -73,7 +73,7 @@ OUString SwDateTimeField::Expand() const
         fVal = GetValue();
 
     if (nOffset)
-        fVal += (double)(nOffset * 60L) / 86400.0;
+        fVal += nOffset * ( 60 / 86400.0 );
 
     return ExpandValue(fVal, GetFormat(), GetLanguage());
 }
